refactor: change the names of grpc ipc to make it clearer what they do
